requirements.txt
# Dependencias del Proyecto - Modelo de Fuga Colsubsidio
# Instalación: pip install -r requirements.txt

# Compatibilidad base y herramientas de build
setuptools>=68.0.0
wheel>=0.41.0
packaging>=23.1

# Data Science Core
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.15.0

# Data Processing
openpyxl>=3.1.2
# xlrd ya no soporta .xlsx, solo .xls
xlrd>=2.0.1

# Configuration Management
pyyaml>=6.0.2
python-dotenv>=1.0.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.23.1
notebook>=6.5.4

# Utilities
tqdm>=4.65.0
joblib>=1.2.0

# Optional: Advanced ML
# xgboost>=1.7.5
# lightgbm>=3.3.5
# shap>=0.41.0
